+++ src/Symfony/Component/JsonStreamer/JsonStreamWriter.php
@@ -18,6 +18,7 @@ use Symfony\Component\JsonStreamer\Mapping\GenericTypePropertyMetadataLoader;
 use Symfony\Component\JsonStreamer\Mapping\PropertyMetadataLoader;
 use Symfony\Component\JsonStreamer\Mapping\PropertyMetadataLoaderInterface;
 use Symfony\Component\JsonStreamer\Mapping\Write\AttributePropertyMetadataLoader;
+use Symfony\Component\JsonStreamer\Transformer\DateIntervalValueObjectTransformer;
 use Symfony\Component\JsonStreamer\Transformer\DateTimeValueObjectTransformer;
 use Symfony\Component\JsonStreamer\Transformer\PropertyValueTransformerInterface;
 use Symfony\Component\JsonStreamer\Transformer\ValueObjectTransformerInterface;
@@ -33,6 +34,7 @@ use Symfony\Component\TypeInfo\TypeResolver\TypeResolver;
  * @psalm-type Options = array{
  *     date_time_format?: string,
  *     date_time_timezone?: string|\DateTimeZone,
+ *     date_interval_format?: string,
  *     include_null_properties?: bool,
  *     ...<string, mixed>,
  * }
@@ -105,6 +107,7 @@ final class JsonStreamWriter implements StreamWriterInterface
         $streamWritersDir ??= sys_get_temp_dir().'/json_streamer/write';
         $transformers += [
             \DateTimeInterface::class => new DateTimeValueObjectTransformer(),
+            \DateInterval::class => new DateIntervalValueObjectTransformer(),
         ];
 
         $transformersContainer = new class($transformers) implements ContainerInterface {