+++ src/Symfony/Component/JsonStreamer/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.1
 ---
 
+ * Add `DateInterval` value object support with `DateIntervalValueObjectTransformer`
  * Add timezone support to `DateTimeValueObjectTransformer`
  * Add value object support
  * Deprecate `DateTimeTypePropertyMetadataLoader` (both Read and Write), date times are handled as value objects